modified:   CMakeLists.txt 	modified:   src/DAQ/CMakeLists.txt 	modified:   src/DAQ/include/ADC.h 	new file:   src/DAQ/include/CLAS12DAQ.h 	modified:   src/DAQ/include/Crate.h 	modified:   src/DAQ/include/CrateModule.h 	modified:   src/DAQ/include/DAQManager.h 	modified:   src/DAQ/include/Discriminator.h 	modified:   src/DAQ/include/FlashADC.h 	new file:   src/DAQ/include/GlobalDistributionCrate.h 	modified:   src/DAQ/include/LinkDef.h 	new file:   src/DAQ/include/Module.h 	modified:   src/DAQ/include/Scaler.h 	new file:   src/DAQ/include/SignalDistribution.h 	new file:   src/DAQ/include/SubSystemProcessor.h 	modified:   src/DAQ/include/Trigger.h 	modified:   src/DAQ/include/TriggerInterface.h 	modified:   src/DAQ/src/ADC.cxx 	new file:   src/DAQ/src/CLAS12DAQ.cxx 	modified:   src/DAQ/src/Crate.cxx 	modified:   src/DAQ/src/Discriminator.cxx 	modified:   src/DAQ/src/FlashADC.cxx 	new file:   src/DAQ/src/GlobalDistributionCrate.cxx 	new file:   src/DAQ/src/Module.cxx 	modified:   src/DAQ/src/Scaler.cxx 	new file:   src/DAQ/src/SignalDistribution.cxx 	new file:   src/DAQ/src/SubSystemProcessor.cxx 	modified:   src/DAQ/src/TDC.cxx 	renamed:    src/DAQ/src/TriggerSupervisor.cxx -> src/DAQ/src/Trigger.cxx 	modified:   src/DAQ/src/TriggerSupervisor.cxx 	modified:   src/Hits/CMakeLists.txt 	new file:   src/Hits/include/CLAS12Event.h 	new file:   src/Hits/include/CLAS12HitsEvent.h 	new file:   src/Hits/include/HTCCHitsEvent.h 	modified:   src/Hits/include/LinkDef.h 	new file:   src/Hits/src/CLAS12Event.cxx 	new file:   src/Hits/src/CLAS12HitsEvent.cxx 	new file:   src/Hits/src/HTCCHitsEvent.cxx 	modified:   src/Sigs/include/DigitizedSignal.h 	modified:   src/Sigs/include/FlashADCSignal.h 	modified:   src/Sigs/src/DigitizedSignal.cxx 	modified:   src/Sigs/src/FlashADCSignal.cxx 	modified:   src/clasdigi/clasdigi_HTCC.cxx 	modified:   src/tests/CMakeLists.txt 	modified:   src/tests/crate_test2.cxx

+++ src/DAQ/include/FlashADC.h
@@ -16,12 +16,12 @@ namespace clas12 {
       class FlashADC : public ModuleChannel {
 
          public : 
-            short     fNSB           = 10;  // Number of samples before TC (<512)
-            short     fNSA           = 20;  // Number of samples after TC  (<512)
-            short     fTriggerWindow = 100; // Trigger window 
+            int     fNSB           = 10;  // Number of samples before TC (<512)
+            int     fNSA           = 20;  // Number of samples after TC  (<512)
+            int     fTriggerWindow = 100; // Trigger window 
 
-            double    fRefTime = 0.0;      // Reference time (eg, trigger)
-            double    fTime    = 0.0;      // Time since last reset 
+            double    fRefTime       = 0.0; // Reference time (eg, trigger)
+            double    fTime          = 0.0; // Time since last reset 
 
             std::vector<int>     fBuffer;  // fADC buffer
             std::vector<int>     fTCs;     // Array of TCs. Stores buffer index
